feat: implement new context menu class and register context menu listener; update actions for improved functionality
