+++ src/Chrome/Utils/actions.ts
@@ -3,4 +3,5 @@ export class Actions {
   static dsr1 = "dsr1";
   static claude = "claude";
   static setResponseIntoTextbox = "setResponseIntoTextbox";
+  static getQuestionPayload = "getQuestionPayload";
 }
